+++ azure-storage-blob/azure/storage/blob/_constants.py
@@ -5,10 +5,10 @@
 # --------------------------------------------------------------------------
 
 __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 
 # x-ms-version for storage service.
-X_MS_VERSION = '2017-04-17'
+X_MS_VERSION = '2017-07-29'
 
 # internal configurations, should not be changed
 _LARGE_BLOB_UPLOAD_MAX_READ_BUFFER_SIZE = 4 * 1024 * 1024