+++ azure-storage-common/azure/storage/common/__init__.py
@@ -14,6 +14,7 @@ from .models import (
     Logging,
     Metrics,
     CorsRule,
+    DeleteRetentionPolicy,
     ServiceProperties,
     AccessPolicy,
     ResourceTypes,