+++ azure-storage-common/azure/storage/common/_error.py
@@ -104,6 +104,8 @@ def _dont_fail_not_exist(error):
 def _http_error_handler(http_error):
     ''' Simple error handler for azure.'''
     message = str(http_error)
+    if 'x-ms-error-code' in http_error.respheader:
+        message += 'ErrorCode: ' + http_error.respheader['x-ms-error-code']
     if http_error.respbody is not None:
         message += '\n' + http_error.respbody.decode('utf-8-sig')
     raise AzureHttpError(message, http_error.status)