+++ azure-storage-file/azure/storage/file/_constants.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------------------------------
 
 __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 
 # x-ms-version for storage service.
-X_MS_VERSION = '2017-04-17'
+X_MS_VERSION = '2017-07-29'